added user caching and better mapping
lots to work around, still more to do, user lookups are better now. group lookups result in a 404, likely a check to determine what this insane API is expecting for a lookup compared to the groupIDs it offers in the original events poll output shows. It's probably a similar tertiary endpoint lookup that will need to be queried. ...yay.